reindex compiler
