+++ src/index.js
@@ -1,13 +1,8 @@
 export * as ast from 'ast';
+export * as compiler from 'compiler';
 
 export Parser from 'parser';
-export Compiler from 'compiler';
 export Plugin from 'plugin';
-
-import JsTargetCompiler from 'compiler/js';
-export var impl = {
-    JsTargetCompiler
-};
 
 import ConstantFolder from 'plugin/ConstantFolder';
 export var plugin = {
